Minor Code refactors and cleanup
