+++ src/main/java/wraith/alloyforgery/recipe/handlers/BlastFurnaceRecipeHandler.java
@@ -1,9 +1,6 @@
 package wraith.alloyforgery.recipe.handlers;
 
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.inventory.Inventory;
-import net.minecraft.inventory.SimpleInventory;
 import net.minecraft.item.ItemStack;
 import net.minecraft.recipe.*;
 import net.minecraft.util.Identifier;
@@ -47,10 +44,6 @@ public class BlastFurnaceRecipeHandler extends ForgeRecipeHandler<BlastingRecipe
         return getFuelPerTick(this.lastRecipe.get()) * context.forgeDefinition().speedMultiplier();
     }
 
-    public static float getFuelPerTick(BlastingRecipe recipe){
-        return ((recipe.getCookTime() / (float) ForgeDefinition.BASE_MAX_SMELT_TIME) * 10);
-    }
-
     @Override
     public void craftRecipe(RecipeContext context, Consumer<DefaultedList<ItemStack>> remainderConsumer) {
         var inputStack = emulatedInv.getStack(0);
@@ -62,74 +55,17 @@ public class BlastFurnaceRecipeHandler extends ForgeRecipeHandler<BlastingRecipe
         super.craftRecipe(context, remainderConsumer);
     }
 
-    private static class SingleSlotInventory implements Inventory {
-
-        public int index = 0;
-        public Inventory wrappedInventory = new SimpleInventory();
-
-        public SingleSlotInventory(){}
-
-        public SingleSlotInventory changeIndex(int index, Inventory wrappedInventory){
-            this.index = index;
-            this.wrappedInventory = wrappedInventory;
-
-            return this;
-        }
-
-        @Override
-        public int size() {
-            return 1;
-        }
-
-        @Override
-        public boolean isEmpty() {
-            return this.wrappedInventory.getStack(index).isEmpty();
-        }
-
-        @Override
-        public ItemStack getStack(int slot) {
-            return this.wrappedInventory.getStack(index);
-        }
-
-        @Override
-        public ItemStack removeStack(int slot, int amount) {
-            return this.wrappedInventory.removeStack(index, amount);
-        }
-
-        @Override
-        public ItemStack removeStack(int slot) {
-            return this.wrappedInventory.removeStack(index);
-        }
-
-        @Override
-        public void setStack(int slot, ItemStack stack) {
-            this.wrappedInventory.setStack(index, stack);
-        }
-
-        @Override
-        public void markDirty() {
-            this.wrappedInventory.markDirty();
-        }
-
-        @Override
-        public boolean canPlayerUse(PlayerEntity player) {
-            return this.wrappedInventory.canPlayerUse(player);
-        }
-
-        @Override
-        public void clear() {}
+    public static float getFuelPerTick(BlastingRecipe recipe){
+        return ((recipe.getCookTime() / (float) ForgeDefinition.BASE_MAX_SMELT_TIME) * 10);
     }
+
+    //----
 
     private static final Set<Identifier> BLACKLIST_BLASTING_RECIPES = new HashSet<>();
 
     public static void initEvents(){
-        ServerLifecycleEvents.SERVER_STARTED.register(server -> {
-            onDatapackload(server.getRecipeManager());
-        });
-
-        ServerLifecycleEvents.END_DATA_PACK_RELOAD.register((server, resourceManager, success) -> {
-            onDatapackload(server.getRecipeManager());
-        });
+        ServerLifecycleEvents.SERVER_STARTED.register(server -> onDatapackload(server.getRecipeManager()));
+        ServerLifecycleEvents.END_DATA_PACK_RELOAD.register((server, manager, success) -> onDatapackload(server.getRecipeManager()));
     }
 
     public static void onDatapackload(RecipeManager manager){
